bot-service/pyproject.toml
[project]
name = "bot-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.8",
    "langchain>=0.3.19",
    "langchain-ai21>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
]

[tool.ruff]
exclude = [
    "migrations/*",
    "app/models/__init__.py",
]
line-length = 120
indent-width = 4
